Remove SerializedSuiteStore
