+++ composer.json
@@ -15,7 +15,7 @@
         "smartassert/results-client": "^9",
         "smartassert/service-client": "^8",
         "smartassert/service-request": "^1.5",
-        "smartassert/sources-client": "^13",
+        "smartassert/sources-client": "13.0",
         "smartassert/users-security-bundle": "^12.0",
         "smartassert/worker-client": "^5",
         "smartassert/worker-manager-client": "^11",
