+++ src/Model/JobStatus.php
@@ -6,6 +6,7 @@ namespace App\Model;
 
 use App\Entity\Machine;
 use App\Entity\ResultsJob;
+use App\Entity\SerializedSuite;
 use App\Enum\JobComponent;
 use App\Services\PreparationStateFactory;
